python-backend/requirements.txt
blinker==1.9.0
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.3
flask-cors==6.0.2
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==11.3.0
pypdf==6.7.5
reportlab==4.4.10
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
